# txmatching/web/swagger/swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
...
        "/v1/user/authentik-login": {
            "get": {
                "responses": {
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    }
                },
                "operationId": "get_authentik_login",
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/change-password": {
            "put": {
                "responses": {
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "200": {
                        "description": "Whether the password was changed successfully.",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    }
                },
                "operationId": "put_password_change_api",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/PasswordChange"
                        }
                    }
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/login": {
            "post": {
                "responses": {
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "200": {
                        "description": "Login successful. JWT generated. User must attach the token to every request in the \"Authorization\" header with the prefix \"Bearer\". Example: \"Authorization: Bearer some_token\", where some_token is the token received in the response.",
                        "schema": {
                            "$ref": "#/definitions/LoginSuccessResponse"
                        }
                    }
                },
                "description": "Endpoint to be used for login of users. Returns JWT. This JWT serves two purposes: in the case of normal accounts it can be used to obtain OTP at the /otp endpoint; in the case of service accounts it can be directly used to access /public/patient-upload endpoint",
                "operationId": "post_login_api",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UserLogin"
                        }
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/otp": {
            "post": {
                "responses": {
                    "503": {
                        "description": "It was not possible to reach the SMS gate, thus the one time password could not be send.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "200": {
                        "description": "OTP validation was successful. JWT generated.",
                        "schema": {
                            "$ref": "#/definitions/LoginSuccessResponse"
                        }
                    }
                },
                "operationId": "post_otp_login_api",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/OtpLogin"
                        }
                    }
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "put": {
                "responses": {
                    "503": {
                        "description": "It was not possible to reach the SMS gate, thus the one time password could not be send.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "200": {
                        "description": "Whether the new OTP was generated and sent.",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    }
                },
                "operationId": "put_otp_login_api",
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/refresh-token": {
            "get": {
                "responses": {
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "200": {
                        "description": "Token successfully refreshed.",
                        "schema": {
                            "$ref": "#/definitions/LoginSuccessResponse"
                        }
                    }
                },
                "operationId": "get_refresh_token_api",
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/register": {
            "post": {
                "responses": {
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "200": {
                        "description": "Detailed info about the registered user.",
                        "schema": {
                            "$ref": "#/definitions/UserRegistrationOut"
                        }
                    }
                },
                "operationId": "post_registration_api",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UserRegistration"
                        }
                    }
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/reset-password": {
            "put": {
                "responses": {
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "200": {
                        "description": "Whether the password reset was successful.",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    }
                },
                "operationId": "put_reset_password",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ResetPassword"
                        }
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/v1/user/{email}/reset-password-token": {
            "parameters": [
                {
                    "name": "email",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "500": {
                        "description": "Unexpected error, see contents for details.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied. You do not have rights to access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication failed.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "400": {
                        "description": "Wrong data format.",
                        "schema": {
                            "$ref": "#/definitions/FailResponse"
                        }
                    },
                    "200": {
                        "description": "Returns reset token.",
                        "schema": {
                            "$ref": "#/definitions/ResetRequestResponse"
                        }
                    }
                },
                "operationId": "get_request_reset",
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        }
    },
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
...
    "definitions": {
...
        "FailResponse": {
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "LoginSuccessResponse": {
            "required": [
                "auth_token"
            ],
            "properties": {
                "auth_token": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "OtpLogin": {
            "required": [
                "otp"
            ],
            "properties": {
                "otp": {
                    "type": "string",
                    "description": "OTP for this login."
                }
            },
            "type": "object"
        },
...
        "PasswordChange": {
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "type": "string",
                    "description": "Current password."
                },
                "new_password": {
                    "type": "string",
                    "description": "New password."
                }
            },
            "type": "object"
        },
...
        "ResetPassword": {
            "required": [
                "password",
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "description": "Reset Token."
                },
                "password": {
                    "type": "string",
                    "description": "New password."
                }
            },
            "type": "object"
        },
        "ResetRequestResponse": {
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "Success": {
            "required": [
                "success"
            ],
            "properties": {
                "success": {
                    "type": "boolean"
                }
            },
            "type": "object"
        },
...
        "UserLogin": {
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "description": "Email of the user to login."
                },
                "password": {
                    "type": "string",
                    "description": "User's password."
                }
            },
            "type": "object"
        },
        "UserRegistration": {
            "required": [
                "allowed_countries",
                "allowed_txm_events",
                "email",
                "password",
                "require_second_factor",
                "role",
                "second_factor"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "description": "Email/username used for authentication."
                },
                "password": {
                    "type": "string",
                    "description": "User's password."
                },
                "role": {
                    "type": "string",
                    "description": "User's role.",
                    "example": "ADMIN",
                    "enum": [
                        "ADMIN",
                        "VIEWER",
                        "EDITOR",
                        "SERVICE"
                    ]
                },
                "require_second_factor": {
                    "type": "boolean",
                    "description": "Whether to require second factor for the user",
                    "example": true
                },
                "second_factor": {
                    "type": "string",
                    "description": "2FA: Phone number for user account in standard format (see example), IP address for SERVICE account.",
                    "example": "+420657123987"
                },
                "allowed_countries": {
                    "type": "array",
                    "description": "Countries that the user has access to.",
                    "example": [
                        "AUT",
                        "CZE"
                    ],
                    "items": {
                        "$ref": "#/definitions/CountryCode"
                    }
                },
                "allowed_txm_events": {
                    "type": "array",
                    "description": "Countries that the user has access to.",
                    "example": [
                        "test"
                    ],
                    "items": {
                        "type": "string",
                        "description": "Name of TXM event"
                    }
                }
            },
            "type": "object"
        },
        "UserRegistrationOut": {
            "required": [
                "allowed_countries",
                "allowed_txm_events",
                "email",
                "password_reset_token",
                "password_reset_token_message",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "description": "Email/username used for authentication."
                },
                "role": {
                    "type": "string",
                    "description": "User's role.",
                    "example": "ADMIN",
                    "enum": [
                        "ADMIN",
                        "VIEWER",
                        "EDITOR",
                        "SERVICE"
                    ]
                },
                "allowed_countries": {
                    "type": "array",
                    "description": "Countries that the user has access to.",
                    "example": [
                        "AUT",
                        "CZE"
                    ],
                    "items": {
                        "$ref": "#/definitions/CountryCode"
                    }
                },
                "allowed_txm_events": {
                    "type": "array",
                    "description": "Countries that the user has access to.",
                    "example": [
                        "test"
                    ],
                    "items": {
                        "type": "string",
                        "description": "Name of TXM event"
                    }
                },
                "password_reset_token": {
                    "type": "string",
                    "description": "token to reset password",
                    "example": "long_string_token"
                },
                "password_reset_token_message": {
                    "type": "string",
                    "description": "detailed description how to reset password including url",
                    "example": "To reset password go tu url: www.reset.password/token"
                }
            },
            "type": "object"
        },
...
    }
}